requirements.txt
streamlit==1.32.0
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
pillow==10.2.0
streamlit-image-coordinates>=0.2.1
